pyproject.toml
[tool.poetry]
name = "lockbox-proxy"
packages = [
    { include = "lockbox" }
]
version = "0.1.2"
description = "Lockbox is a forward proxy for making third party API calls."
authors = ["Jackie Tung <jackie.tung@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/mkjt2/lockbox"
repository = "https://github.com/mkjt2/lockbox"

[tool.poetry.dependencies]
python = ">=3.10"
flask = "^3.0.1"
types-flask = "^1.1.6"
requests = "^2.31.0"
gunicorn = "^21.2.0"
pydantic = "^2.5.3"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.348"
pre-commit = "^3.6.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
# TODO not portable
venvPath = "."
venv = "venv"
